+++ internal/pkg/user/repository.go
@@ -5,16 +5,16 @@ import (
 )
 
 type RepoUser interface {
-	GetUserByIDWithPass(ID int) (models.User, error)
-	GetUserByIDWithoutPass(ID int) (models.User, error)
+	GetUserByIDWithPass(ID int64) (models.User, error)
+	GetUserByIDWithoutPass(ID int64) (models.User, error)
 	GetUserByLoginWithPass(login string) (models.User, error)
 	GetByLoginWithoutPass(login string) (models.User, error)
-	GetLoginByID(ID int) (string, error)
-	GetIDByLogin(login string) (int, error)
+	GetLoginByID(ID int64) (string, error)
+	GetIDByLogin(login string) (int64, error)
 	Create(newUser models.User) error
 	Update(usrUpd models.User) error
 	IsExists(user models.User) (bool, error)
-	DeleteByID(ID int) error
+	DeleteByID(ID int64) error
 	CheckPass(login string, newpass string) (bool, error)
 	UploadAvatar(Name string, Content []byte) error
 	UpdateAvatarPath(User models.User, Name string) error
